funcianalidad de compatibilidad finalizada

+++ application/controllers/almacen/Productos.php
@@ -46,19 +46,16 @@ class Productos extends CI_Controller {
 	public function store(){
 
 		$codigo_barras = $this->input->post("codigo_barras");
-		$year_id = $this->input->post("year_id");
-		$fabricante_id = $this->input->post("fabricante_id");
-		$modelo_id = $this->input->post("modelo_id");
 		$calidad_id = $this->input->post("calidad_id");
 		$nombre = $this->input->post("nombre");
-		$categoria_id = $this->input->post("categoria_id");
-		$subcategoria_id = $this->input->post("subcategoria_id");
 		$descripcion = $this->input->post("descripcion");
-		$marca_id = $this->input->post("marca_id");
-		$presentacion_id = $this->input->post("presentacion_id");
 		$stock_minimo = $this->input->post("stock_minimo");
 
+		$marcas = $this->input->post("marcas");
 		$modelos = $this->input->post("modelos");
+		$range_year = $this->input->post("range_year");
+		$year_until = $this->input->post("year_until");
+		$year_from = $this->input->post("year_from");
 		$idProductosA = $this->input->post("idProductosA");
 		$cantidadesA = $this->input->post("cantidadesA");
 		$idPrecios = $this->input->post("idPrecios");
@@ -68,39 +65,33 @@ class Productos extends CI_Controller {
 		$this->form_validation->set_rules("nombre","Nombre","required|is_unique[productos.nombre]");
 
 		if ($this->form_validation->run()==TRUE) {
-			$imagen = 'image_default.jpg';
-			if (!empty($_FILES['imagen']['name'])) {
-				$config['upload_path']          = './assets/imagenes_productos/';
-                $config['allowed_types']        = 'gif|jpg|png';
-
-                $this->load->library('upload', $config);
-                if ($this->upload->do_upload('imagen'))
-                {
-  					$data = array('upload_data' => $this->upload->data());
-                    $imagen = $data['upload_data']['file_name'];
-                } 
-			}
 			$data  = array(
 				"nombre" => $nombre, 
 				"descripcion" => $descripcion,
-				"fabricante_id" => $fabricante_id,
-				"year_id" => $year_id,
-				"marca_id" => $marca_id,
-				"modelo_id" => $modelo_id,
 				"calidad_id" => $calidad_id,
-				"presentacion_id" => $presentacion_id,
 				"stock_minimo" => $stock_minimo,
-				"categoria_id" => $categoria_id,
-				"subcategoria_id" => $subcategoria_id,
 				"codigo_barras" => $codigo_barras,
-				"imagen" => $imagen,
 				"estado" => "1"
 			);
 			$producto = $this->Comun_model->insert("productos", $data);
 			if ($producto) {
+				$imagen = 'image_default.jpg';
+				if (!empty($_FILES['imagen']['name'])) {
+					$config['upload_path']          = './assets/imagenes_productos/';
+	                $config['allowed_types']        = 'gif|jpg|png';
+
+	                $this->load->library('upload', $config);
+	                if ($this->upload->do_upload('imagen'))
+	                {
+	  					$data = array('upload_data' => $this->upload->data());
+	                    $imagen = $data['upload_data']['file_name'];
+	                } 
+				}
+				$data  = array('imagen' => $imagen);
+				$this->Comun_model->update("productos","id=".$producto->id,$data);
 				$this->generateBarCode($codigo_barras);
 				if (!empty($modelos)) {
-					$this->saveCompatibilidades($producto->id,$modelos);
+					$this->saveCompatibilidades($producto->id,$modelos,$marcas,$range_year,$year_from,$year_until);
 				}
 				if (!empty($idProductosA)) {
 					$this->saveAsociados($producto->id,$idProductosA,$cantidadesA);
@@ -133,11 +124,25 @@ class Productos extends CI_Controller {
 		}
 	}
 
-	protected function saveCompatibilidades($producto_id,$modelos){
+	protected function saveCompatibilidades($producto_id,$modelos,$marcas,$range_year,$year_from,$year_until){
+
 		for ($i=0; $i < count($modelos) ; $i++) { 
+			$concat_year = "";
+			if ($range_year[$i]) {
+				for ($j=$year_from[$i]; $j <= $year_until[$i] ; $j++) { 
+					$concat_year .= $j.",";
+				}
+			}else{
+				$concat_year = $year_from[$i];
+			}
 			$data = array(
 				'producto_id' => $producto_id, 
 				'modelo_id' => $modelos[$i],
+				'marca_id' => $marcas[$i],
+				'range_year' => $range_year[$i],
+				'year_from' => $year_from[$i],
+				'year_until' => $year_until[$i],
+				'concat_year' => $concat_year,
 			);
 			$this->Comun_model->insert("compatibilidades",$data);
 		}
@@ -159,13 +164,7 @@ class Productos extends CI_Controller {
 		$contenido_interno  = array(
 			//"permisos" => $this->permisos,
 			"producto" => $producto,
-			"categorias" => $this->Comun_model->get_records("categorias","estado=1"),
-			"subcategorias" => $this->Comun_model->get_records("subcategorias","categoria_id='$producto->categoria_id' and estado=1"),
-			"years" => $this->Comun_model->get_records("years","estado=1"), 
-			"presentaciones" => $this->Comun_model->get_records("presentaciones","estado=1"), 
-			"modelos" => $this->Comun_model->get_records("modelos","estado=1"), 
 			"marcas" => $this->Comun_model->get_records("marcas","estado=1"), 
-			"fabricantes" => $this->Comun_model->get_records("fabricantes","estado=1"), 
 			"calidades" => $this->Comun_model->get_records("calidades","estado=1"), 
 			"tipo_precios" => $this->Comun_model->get_records("precios","estado=1"), 
 			"compatibilidades" => $this->Comun_model->get_records("compatibilidades","producto_id='$id'"),
@@ -183,19 +182,16 @@ class Productos extends CI_Controller {
 	public function update(){
 		$idProducto = $this->input->post("idProducto");
 		$codigo_barras = $this->input->post("codigo_barras");
-		$year_id = $this->input->post("year_id");
-		$fabricante_id = $this->input->post("fabricante_id");
-		$modelo_id = $this->input->post("modelo_id");
 		$calidad_id = $this->input->post("calidad_id");
 		$nombre = $this->input->post("nombre");
-		$categoria_id = $this->input->post("categoria_id");
-		$subcategoria_id = $this->input->post("subcategoria_id");
 		$descripcion = $this->input->post("descripcion");
-		$marca_id = $this->input->post("marca_id");
-		$presentacion_id = $this->input->post("presentacion_id");
 		$stock_minimo = $this->input->post("stock_minimo");
 
+		$marcas = $this->input->post("marcas");
 		$modelos = $this->input->post("modelos");
+		$range_year = $this->input->post("range_year");
+		$year_until = $this->input->post("year_until");
+		$year_from = $this->input->post("year_from");
 		$idProductosA = $this->input->post("idProductosA");
 		$cantidadesA = $this->input->post("cantidadesA");
 		$idPrecios = $this->input->post("idPrecios");
@@ -214,37 +210,35 @@ class Productos extends CI_Controller {
 
 		if ($this->form_validation->run()==TRUE) {
 			$producto = $this->Comun_model->get_record("productos","id='$idProducto'");
-			$imagen = $producto->imagen;
-			if (!empty($_FILES['imagen']['name'])) {
-				$config['upload_path']          = './assets/imagenes_productos/';
-                $config['allowed_types']        = 'gif|jpg|png';
-
-                $this->load->library('upload', $config);
-                if ($this->upload->do_upload('imagen'))
-                {
-  					$data = array('upload_data' => $this->upload->data());
-                    $imagen = $data['upload_data']['file_name'];
-                    unlink("assets/imagenes_productos/".$producto->imagen);
-                } 
-			}
+			
 			$data = array(
 				"nombre" => $nombre,
 				"descripcion" => $descripcion,
-				"fabricante_id" => $fabricante_id,
-				"year_id" => $year_id,
-				"marca_id" => $marca_id,
-				"modelo_id" => $modelo_id,
 				"calidad_id" => $calidad_id,
-				"presentacion_id" => $presentacion_id,
 				"stock_minimo" => $stock_minimo,
-				"categoria_id" => $categoria_id,
-				"subcategoria_id" => $subcategoria_id,
 				"codigo_barras" => $codigo_barras,
-				"imagen" => $imagen
 			);
 
 			if ($this->Comun_model->update("productos","id=$idProducto",$data)) {
+				$imagen = $producto->imagen;
+				if (!empty($_FILES['imagen']['name'])) {
+					$config['upload_path']          = './assets/imagenes_productos/';
+	                $config['allowed_types']        = 'gif|jpg|png';
 
+	                $this->load->library('upload', $config);
+	                if ($this->upload->do_upload('imagen'))
+	                {
+	  					$data = array('upload_data' => $this->upload->data());
+	                    $imagen = $data['upload_data']['file_name'];
+	                    if ($producto->imagen != "image_default.jpg") {
+	                    	unlink("assets/imagenes_productos/".$producto->imagen);
+	                    }
+	                    
+	                } 
+				}
+
+				$data  = array('imagen' => $imagen);
+				$this->Comun_model->update("productos","id=$idProducto",$data);
 				$this->generateBarCode($codigo_barras);
 
 				$this->Comun_model->delete("compatibilidades","producto_id='$idProducto'");
@@ -253,7 +247,7 @@ class Productos extends CI_Controller {
 				$this->Comun_model->delete("producto_precio","producto_id='$idProducto'");
 
 				if (!empty($modelos)) {
-					$this->saveCompatibilidades($idProducto,$modelos);
+					$this->saveCompatibilidades($producto->id,$modelos,$marcas,$range_year,$year_from,$year_until);
 				}
 				if (!empty($idProductosA)) {
 					$this->saveAsociados($idProducto,$idProductosA,$cantidadesA);
